Adding login container
