+++ src/main/webapp/login/loginModal.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import Dialog from '@material-ui/core/Dialog'
-import LogIn from './login'
+import Container from './container'
 
 export const LogInModal = props => {
   const { handleClose, open, label } = props
   return (
     <Dialog open={open} onClose={handleClose}>
-      <LogIn handleClose={handleClose} label={label} />
+      <Container handleClose={handleClose} label={label} />
     </Dialog>
   )
 }
